use std::uint64_t for all node ID variables. Hope, I found them all…

+++ scriptengine/decorators/segmentlistdecorator.h
@@ -16,8 +16,8 @@ signals:
 public slots:
     nodeListElement * source(segmentListElement *self);
     nodeListElement * target(segmentListElement *self);
-    int source_id(segmentListElement *self);
-    int target_id(segmentListElement *self);
+    uint64_t source_id(segmentListElement *self);
+    uint64_t target_id(segmentListElement *self);
     QString static_Segment_help();
 };
